Check for implicit relative imports.
It's kind of heuristic-y and I can think of at least one case where it
might be considered a false positive, but I think it's good enough.

For posterity, that one case is a setup.py that imports from the
adjacent source package, but I'm not a fan of that approach anyway.