Initial pass at handling long lines better.
After running into issues with 79 columns being good for some
situations, but less good for others, I finally decided to just write my
own long line lint thing. It's maybe a bit too heuristic, but we'll see
how well it works out in practice.

Basically, it allows lines to be 79 < length <= 119, but only if
two-thirds of the line are comments or string literals by character
count.

Committing this to run it on CI and get feedback on the docs before
merging to master.